requirements.txt
discord.py
sqlite-utils
datasette
sentence_transformers
python-dotenv
openai
click